CalculationsScreen.kt + вынес навигацию в отдельные классы
