+++ app/src/main/java/com/mvptest/presentation/ui/bottommenu/NavigationItem.kt
@@ -6,5 +6,4 @@ sealed class NavigationItem(var route: String, var icon: Int) {
     object Home : NavigationItem("home", R.drawable.ic_home)
     object Calculating : NavigationItem("calculating", R.drawable.ic_calculator)
     object MyProjects : NavigationItem("projects_list", R.drawable.ic_notes)
-    object Profile : NavigationItem("account", R.drawable.ic_profile)
 }
