+++ app/src/main/java/com/mvptest/presentation/ui/calculation/NavigationCalculation.kt
@@ -2,6 +2,7 @@ package com.mvptest.presentation.ui.calculation
 
 import androidx.compose.runtime.Composable
 import androidx.navigation.NavController
+import androidx.navigation.NavGraph
 import androidx.navigation.NavGraphBuilder
 import androidx.navigation.compose.composable
 import androidx.navigation.navigation
@@ -12,6 +13,7 @@ import com.mvptest.presentation.ui.calculation.calculators.AirHeater
 import com.mvptest.presentation.ui.calculation.calculators.Conditioner
 import com.mvptest.presentation.ui.calculation.calculators.Diffusers
 import com.mvptest.presentation.ui.calculation.calculators.DuctCrossSection
+import com.mvptest.presentation.ui.project.newproject.NavigationNewProjectItem
 
 sealed class NavigationCalculationItem(val route: String) {
     object AirExchange : NavigationCalculationItem(route = "air_exchange")
@@ -23,9 +25,22 @@ sealed class NavigationCalculationItem(val route: String) {
 }
 
 fun NavGraphBuilder.calculationGraph(navController: NavController) {
-    navigation(startDestination = "calculating", route = NavigationItem.Calculating.route) {
+    navigation(startDestination = NavigationItem.Calculating.route, route = "calculation") {
+        composable(NavigationItem.Calculating.route) {
+            CalculatingMainScreen(onItemClicked = { route ->
+                navController.navigate(route)
+            })
+        }
         composable(NavigationCalculationItem.AirExchange.route) {
-            AirExchangeScreen()
+            AirExchangeScreen(onBackPressed = {
+                navController.navigate(NavigationItem.Calculating.route) {
+                    popUpTo(
+                        NavigationItem.Calculating.route
+                    ) {
+                        inclusive = true
+                    }
+                }
+            })
         }
         composable(NavigationCalculationItem.DuctCrossSection.route) {
             DuctCrossSection()
