+++ app/src/main/java/com/mvptest/presentation/ui/calculation/calculators/AirExchangeScreen.kt
@@ -1,8 +1,17 @@
 package com.mvptest.presentation.ui.calculation.calculators
 
+import androidx.activity.compose.BackHandler
 import androidx.compose.runtime.Composable
 
 @Composable
-fun AirExchangeScreen(){
+fun AirExchangeScreen(onBackPressed: () -> Unit){
 
+
+
+    @Composable
+    fun TestScreen() {
+        BackHandler {
+            onBackPressed()
+        }
+    }
 }
