--- app/src/main/java/com/mvptest/presentation/ui/project/newproject/NavigationNewProjectItem.kt
@@ -1,7 +1,0 @@
-package com.mvptest.presentation.ui.project.newproject
-
-sealed class NavigationNewProjectItem(var route: String) {
-    object First : NavigationNewProjectItem("new_project_first")
-    object Second : NavigationNewProjectItem("new_project_second")
-    object ProjectDetails : NavigationNewProjectItem("project_details/{projectId}")
-}
